generate-pacreport: improve signal handling
For shells that aren't bash the EXIT trap handler will not run if the
script is interrupted.  It is only evaluated at the end of a script upon
normal termination, as if the code to be executed was added at the end
of a script.

This patch adds handlers for HUP, INT, and TERM which represent a common
set of signals this script may see.  It resends the signal as we intend
to terminate on them as would normally happen, but with the additional
temp file removal.

This also combines the two find commands into one avoiding the need to
run it twice.