GNUmakefile: rename and fix Makefile
As the makefile depends on GNU make features the file should be called
GNUmakefile, which will also benefit systems that don't use GNU make
per-default and would have to otherwise call gmake.

This corrects the need for tabs as separators for recipe commands.

In addition I've renamed the macros to be more in line with GNU's
conventions[0] while also adding support for DESTDIR usage.

1. https://www.gnu.org/prep/standards/html_node/Directory-Variables.html